Initial commit of darkenergycomponent trait class
