+++ astropy/cosmology/_src/flrw/base.py
@@ -4,12 +4,11 @@ __all__ = ["FLRW", "FlatFLRWMixin"]
 
 import inspect
 import warnings
-from abc import abstractmethod
 from collections.abc import Mapping
 from dataclasses import field
 from functools import cached_property
 from inspect import signature
-from math import exp, floor, log, pi, sqrt
+from math import floor, pi, sqrt
 from numbers import Number
 from typing import Self, TypeVar, overload
 
@@ -34,6 +33,7 @@ from astropy.cosmology._src.parameter import (
     validate_with_unit,
 )
 from astropy.cosmology._src.traits import (
+    DarkEnergyComponent,
     HubbleParameter,
     ScaleFactor,
     TemperatureCMB,
@@ -93,10 +93,11 @@ class FLRW(
     Cosmology,
     HubbleParameter,
     ScaleFactor,
-    TemperatureCMB,
     _CriticalDensity,
     _BaryonComponent,
     _MatterComponent,
+    DarkEnergyComponent,
+    TemperatureCMB,
 ):
     """An isotropic and homogeneous (Friedmann-Lemaitre-Robertson-Walker) cosmology.
 
@@ -163,7 +164,10 @@ class FLRW(
         doc="Omega matter; matter density/critical density at z=0.",
         fvalidate="non-negative",
     )
-    Ode0: Parameter = ParameterOde0.clone()
+    Ode0: Parameter = Parameter(
+        doc="Omega dark energy; dark energy density/critical density at z=0.",
+        fvalidate="float",
+    )
     Tcmb0: Parameter = Parameter(
         default=0.0 * u.K,
         doc="Temperature of the CMB at z=0.",
@@ -357,36 +361,6 @@ class FLRW(
 
     # ---------------------------------------------------------------
 
-    @abstractmethod
-    @deprecated_keywords("z", since="7.0")
-    def w(self, z):
-        r"""The dark energy equation of state.
-
-        Parameters
-        ----------
-        z : Quantity-like ['redshift'], array-like
-            Input redshift.
-
-            .. versionchanged:: 7.0
-                Passing z as a keyword argument is deprecated.
-
-        Returns
-        -------
-        w : ndarray or float
-            The dark energy equation of state.
-            `float` if scalar input.
-
-        Notes
-        -----
-        The dark energy equation of state is defined as
-        :math:`w(z) = P(z)/\rho(z)`, where :math:`P(z)` is the pressure at
-        redshift z and :math:`\rho(z)` is the density at redshift z, both in
-        units where c=1.
-
-        This must be overridden by subclasses.
-        """
-        raise NotImplementedError("w(z) is not implemented")
-
     @deprecated_keywords("z", since="7.0")
     def Otot(self, z):
         """The total density parameter at redshift ``z``.
@@ -456,30 +430,6 @@ class FLRW(
         if self.Ok0 == 0:  # Common enough to be worth checking explicitly
             return np.zeros(z.shape) if hasattr(z, "shape") else 0.0
         return self.Ok0 * (z + 1.0) ** 2 * self.inv_efunc(z) ** 2
-
-    @deprecated_keywords("z", since="7.0")
-    def Ode(self, z):
-        """Return the density parameter for dark energy at redshift ``z``.
-
-        Parameters
-        ----------
-        z : Quantity-like ['redshift'], array-like
-            Input redshift.
-
-            .. versionchanged:: 7.0
-                Passing z as a keyword argument is deprecated.
-
-        Returns
-        -------
-        Ode : ndarray or float
-            The density of non-relativistic matter relative to the critical
-            density at each redshift.
-            Returns `float` if the input is scalar.
-        """
-        z = aszarr(z)
-        if self.Ode0 == 0:  # Common enough to be worth checking explicitly
-            return np.zeros(z.shape) if hasattr(z, "shape") else 0.0
-        return self.Ode0 * self.de_density_scale(z) * self.inv_efunc(z) ** 2
 
     @deprecated_keywords("z", since="7.0")
     def Ogamma(self, z):
@@ -620,80 +570,6 @@ class FLRW(
         rel_mass = rel_mass_per.sum(-1) + self._nmasslessnu
 
         return prefac * self._neff_per_nu * rel_mass
-
-    def _w_integrand(self, ln1pz, /):
-        """Internal convenience function for w(z) integral (eq. 5 of [1]_).
-
-        Parameters
-        ----------
-        ln1pz : `~numbers.Number` or scalar ndarray, positional-only
-            Assumes scalar input, since this should only be called inside an
-            integral.
-
-            .. versionchanged:: 7.0
-                The argument is positional-only.
-
-        References
-        ----------
-        .. [1] Linder, E. (2003). Exploring the Expansion History of the
-               Universe. Phys. Rev. Lett., 90, 091301.
-        """
-        return 1.0 + self.w(exp(ln1pz) - 1.0)
-
-    @deprecated_keywords("z", since="7.0")
-    def de_density_scale(self, z):
-        r"""Evaluates the redshift dependence of the dark energy density.
-
-        Parameters
-        ----------
-        z : Quantity-like ['redshift'], array-like
-            Input redshift.
-
-            .. versionchanged:: 7.0
-                Passing z as a keyword argument is deprecated.
-
-        Returns
-        -------
-        I : ndarray or float
-            The scaling of the energy density of dark energy with redshift.
-            Returns `float` if the input is scalar.
-
-        Notes
-        -----
-        The scaling factor, I, is defined by :math:`\rho(z) = \rho_0 I`,
-        and is given by
-
-        .. math::
-
-           I = \exp \left( 3 \int_{a}^1 \frac{ da^{\prime} }{ a^{\prime} }
-                          \left[ 1 + w\left( a^{\prime} \right) \right] \right)
-
-        The actual integral used is rewritten from [1]_ to be in terms of z.
-
-        It will generally helpful for subclasses to overload this method if
-        the integral can be done analytically for the particular dark
-        energy equation of state that they implement.
-
-        References
-        ----------
-        .. [1] Linder, E. (2003). Exploring the Expansion History of the
-               Universe. Phys. Rev. Lett., 90, 091301.
-        """
-        # This allows for an arbitrary w(z) following eq (5) of
-        # Linder 2003, PRL 90, 91301.  The code here evaluates
-        # the integral numerically.  However, most popular
-        # forms of w(z) are designed to make this integral analytic,
-        # so it is probably a good idea for subclasses to overload this
-        # method if an analytic form is available.
-        z = aszarr(z)
-        if not isinstance(z, (Number, np.generic)):  # array/Quantity
-            ival = np.array(
-                [quad(self._w_integrand, 0, log(1 + redshift))[0] for redshift in z]
-            )
-            return np.exp(3 * ival)
-        else:  # scalar
-            ival = quad(self._w_integrand, 0, log(z + 1.0))[0]
-            return exp(3 * ival)
 
     @deprecated_keywords("z", since="7.0")
     def efunc(self, z):
