+++ astropy/cosmology/_src/traits/__init__.py
@@ -5,6 +5,7 @@ The public API is provided by `astropy.cosmology.traits`.
 """
 
 __all__ = [
+    "DarkEnergyComponent",
     "HubbleParameter",
     "ScaleFactor",
     "TemperatureCMB",
@@ -15,6 +16,7 @@ __all__ = [
 
 from ._matter_density import _MatterComponent
 from .baryons import _BaryonComponent
+from .darkenergy import DarkEnergyComponent
 from .hubble import HubbleParameter
 from .rhocrit import _CriticalDensity
 from .scale_factor import ScaleFactor
